+++ src/main.hip.cpp
@@ -16,7 +16,7 @@ extern "C" {
 #define BLOCK_Y 32
 
 
-__constant__ __device__ float d_filter[25];
+__constant__ __device__ int d_filter[25];
 // Apply convolutional filter on image data
 __global__
 void applyFilter(pixel *out, pixel *in, unsigned int width, unsigned int height, unsigned int filterDim, float filterFactor) {
